moved the implementation class for column types to the column package (and out of the API package)
